refactor: moved directory-item to directory-items
