+++ src/directory-items/directory-item.ts
