+++ src/file/file.ts
@@ -11,7 +11,7 @@ import { downloadData, generateBlockName } from './handler'
 import { blocksToManifest, getFileMetadataRawBytes } from './adapter'
 import { Blocks, DataUploadOptions } from './types'
 import { Data } from '@ethersphere/bee-js'
-import { addEntryToDirectory } from '../common/handler'
+import { addEntryToDirectory } from '../directory-items/handler'
 
 /**
  * Files management class
